diarization-service/requirements.txt
fastapi
uvicorn
redis
pyannote.audio
python-dotenv
pydantic
pydantic-settings
torch